Skill Animation & Skill Damage Setting
1. 스킬 시전 모션과 시전 위치, 애니메이션 완료

2. 스킬 데미지 설정. 컴벳시스템과 연동

+++ Assets/3.Script/Monster/Goblin.cs
@@ -12,8 +12,8 @@ public class Goblin : MonoBehaviour, IDamageAble
     public class MonsterStat
     {
         public string name;
-        public int hp;
-        public int maxHp;
+        public float hp;
+        public float maxHp;
         public float range;
         public float speed;
     }
